Move the Permission class to it's own model in the permission directory.
